Feature/breakout card uploader (#7)
* npm audit fixes (#6)


* break out card uploader
