+++ package.json
@@ -17,7 +17,8 @@
   },
   "keywords": [
     "cedh",
-    "codex"
+    "codex",
+    "foundry"
   ],
   "author": "Josh Davidson",
   "license": "MIT",
